Created 2 separate component for navigation
