+++ resources/assets/js/app/components/MainNavigation/index.js
@@ -1,0 +1,1 @@
+export default from './MainNavigation';
